Move kernel_filesystems functionality into a little library.
This way, tasks can know what the kernel filesystems are if they need to.
Move it up a directory, so tasks and formats can share.
